% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignReviewers.R
\name{assignReviewers}
\alias{assignReviewers}
\title{A function to generate an incomplete block experimental design that pairs applicants with reviewers}
\usage{
assignReviewers(
  number.alternatives,
  number.blocks,
  alternatives.per.block = 14,
  n.repeats = 1,
  nReps = 300
)
}
\arguments{
\item{number.alternatives}{Number of applicants}

\item{number.blocks}{Number of reviewers}

\item{alternatives.per.block}{The desired number of applicants that will be assigned to each reviewer (default = 14)}

\item{n.repeats}{Number of times algDesign::optBlock() is repeated (default = 1)}

\item{nReps}{Number of replicates performed within algDesign::optBlock (default = 300)}
}
\value{
A list with 6 components, one of which is the "design." The design has two components, blockID, which refers to row number of that reviewer in the source data, and alternative, which referes to the number of applicants to be reviewed.
}
\description{
A function to generate an incomplete block experimental design that pairs applicants with reviewers
}
\examples{
\dontshow{
sample.design <- assignReviewers(number.alternatives = 30, 
                                 number.blocks = 20,
                                 nReps = 10)


}
\dontrun{

sample.design <- assignReviewers(number.alternatives = n.applics, 
                                 number.blocks = n.sr)

}
}
