% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPI.R
\name{checkPI}
\alias{checkPI}
\title{Identifying missing persons}
\usage{
checkPI(reviewersDF, applicantsDF)
}
\arguments{
\item{reviewersDF}{A data.frame() containing reviewers}

\item{applicantsDF}{A data.frame() containing applicants' faculty of interest, in columns "Faculty.Member.1" through "Faculty.Member.6"}
}
\value{
If all faculty reviewer names match applicants' listed faculty of interest, returns a message indicating that the names match and returns `NULL`. Otherwise, prints a message indicating that some names are not present in the reviewer dataset and a vector of faculty names from the `applicantsDF` that are not present in `reviewersDF`.
}
\description{
Identifying missing persons
}
\examples{
\dontrun{
missingFaculty = checkPI(reviewersDF = reviewers, applicantsDF = applicants)
}
}
