% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineOutputs.R
\name{combineOutputs}
\alias{combineOutputs}
\title{A function to compile assignments from multiple iterations of `incomplete.block.design()`}
\usage{
combineOutputs(reviewers, assignments1, assignments2)
}
\arguments{
\item{reviewers}{A data.frame() with columns `Name` ("Lastname, Firstname"), `Email`, and `type` (type indicates "Student" or "Faculty")}

\item{assignments1}{A matrix (as generated by `getAssignments()`) containing applicant assignments for one reviewer pool (e.g. "Students")}

\item{assignments2}{A matrix (as generated by `getAssignments()`) containing applicant assignments for a second reviewer pool (e.g. "Faculty")}
}
\value{
A data.frame() containing one row per reviewer and "Name", "Email" and "type" columns that indicate reviewer identity, and then "Applicant-1" through "Applicant-i" columns where i is the number of applicants assigned to a single reviewer as defined by `alternatives.per.block` in `assignReviewers()`.
}
\description{
A function to compile assignments from multiple iterations of `incomplete.block.design()`
}
\examples{
\dontrun{
complete.design = combineOutputs(reviewers = reviewers,
assignments1 = s.des.names,
assignments2 = f.des.names)
}
}
