% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssignments.R
\name{getAssignments}
\alias{getAssignments}
\title{Functions to identify reviewer-applicant pairs from incomplete block design}
\usage{
getAssignments(x, appNames)
}
\arguments{
\item{x}{A design generated by `assignReviewers()`}

\item{appNames}{A vector of applicant names}
}
\value{
A matrix of applicant assignments. Each row corresponds with the applicants assigned to each reviewer, so the number of rows should correspond with the number of reviewers. The number of columns corresponds to the parameter `alternatives.per.block` used in `assignReviewers()`.
}
\description{
Functions to identify reviewer-applicant pairs from incomplete block design
}
\examples{
\dontrun{
s.des.names = getAssignments(x = student.design,
                             appNames = applicantNames)
}
}
