% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameFinder.R
\name{nameFinder}
\alias{nameFinder}
\title{A function to extract names from applicant info using the sampling design output from incomplete.block.design(). This is an internal function and users will not directly interact with it.}
\usage{
nameFinder(x, applNames)
}
\arguments{
\item{x}{A single element in the design generated by `assignReviewers()`}

\item{applNames}{A vector of applicant names}
}
\value{
A vector of applicant assignments.
}
\description{
A function to extract names from applicant info using the sampling design output from incomplete.block.design(). This is an internal function and users will not directly interact with it.
}
\examples{
\dontrun{
outNames = nameFinder(x, applNames)
}
}
