% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeConflicts.R
\name{summarizeConflicts}
\alias{summarizeConflicts}
\title{A function to identify where conflicts of interest exist}
\usage{
summarizeConflicts(COIs, ASSIGNs)
}
\arguments{
\item{COIs}{A data.frame() containing conflicts of interest with columns for applicant's "Name" (format "Lastname, Firstname") and "Faculty.Member.1" through "Faculty.Member.6"}

\item{ASSIGNs}{The complete reviewer assignments dataset (as generated by combineOutputs)}
}
\value{
A data.frame() containing one row per reviewer and "Name", "Email" and "Type" columns that indicate reviewer identity, and then "Applicant-1" through "Applicant-i" columns where i is the number of applicants assigned to a single reviewer as defined by `alternatives.per.block` in `assignReviewers()`.
}
\description{
A function to identify where conflicts of interest exist
}
\examples{
\dontrun{
complete.design = combineOutputs(reviewers = reviewers,
assignments1 = s.des.names,
assignments2 = f.des.names)
}
}
